製作程式/建構作業系統核心.sh - 修正子模組未初始化時 clean_up 仍試圖重設子模組的問題

+++ 製作程式/建構作業系統核心.sh
@@ -34,25 +34,27 @@ clean_up() {
 		rmdir "$workaround_safe_build_directory" 2>&1 | tee --append "$PROJECT_LOGS_DIRECTORY/建構作業系統核心.log"
 	fi
 	
-	cd "$PROJECT_THIRD_PARTY_LINUX_SOURCE_DIRECTORY"
-	rm -rf build 2>&1 | tee --append "$PROJECT_LOGS_DIRECTORY/建構作業系統核心.log"
-	# reset tracked files
-	git reset --hard 2>&1 | tee --append "$PROJECT_LOGS_DIRECTORY/建構作業系統核心.log"
-	# remove untracked files
-	git clean --force -d -x 2>&1 | tee --append "$PROJECT_LOGS_DIRECTORY/建構作業系統核心.log"
-	# checkout back to master branch
-	git checkout master 2>&1 | tee --append "$PROJECT_LOGS_DIRECTORY/建構作業系統核心.log"
+	if [ -f "$PROJECT_THIRD_PARTY_LINUX_SOURCE_DIRECTORY/.git" ]; then
+		cd "$PROJECT_THIRD_PARTY_LINUX_SOURCE_DIRECTORY"
+		rm -rf build 2>&1 | tee --append "$PROJECT_LOGS_DIRECTORY/建構作業系統核心.log"
+		# reset tracked files
+		git reset --hard 2>&1 | tee --append "$PROJECT_LOGS_DIRECTORY/建構作業系統核心.log"
+		# remove untracked files
+		git clean --force -d -x 2>&1 | tee --append "$PROJECT_LOGS_DIRECTORY/建構作業系統核心.log"
+		# checkout back to master branch
+		git checkout master 2>&1 | tee --append "$PROJECT_LOGS_DIRECTORY/建構作業系統核心.log"
+	fi
 }
 
 # Defensive Bash Programming - main function, program entry point
 # http://www.kfirlavi.com/blog/2012/11/14/defensive-bash-programming/
 main() {
+	# 預防程式先前被強制終止我們在開始之前多做一次清潔程序
+	clean_up
+
 	# 更新 Git 子模組
 	git submodule init
 	git submodule update --force --depth 1
-	
-	# 預防程式先前被強制終止我們在開始之前多做一次清潔程序
-	clean_up
 
 	# 將來源碼切換到我們要用的版本
 	cd "$PROJECT_THIRD_PARTY_LINUX_SOURCE_DIRECTORY"
